Build require pear1's dependencies to avoid mismatched php stacks
Yum is easily confused by the IUS approach of multiple packages being
able to satisfy pear1's dependencies.  This isn't currently a problem
for php74, because yum will resolve the dependency to the highest
package version that satisifies the dependency.  It will become a
problem as soon as we have a newer version of php.