Remove -n flag from %setup
This is normally appropriate for IUS packages, but in the case of pecl
modules, it is unnecessary.  Pecl archives are "tarbombs" that have top
level files outside any directory.  To have all files accessible during
the build, standard practice is to use %setup's -c flag, which creates
the working directory before extracting the archive.  This is directory
defaults to %{name}-%{version}, but because we're creating it ourselves
and not relying on it matching the archive's top level directory, it can
be named anything.