Correctly register pecl extension with the cli tool
Fedora installs the module xml file with the package name, but since we
modify the package name in IUS it is more accurate to use just the
upstream module name.  Fedora also has file triggers that handle
registering files installed to %{pecl_xmldir} with the pecl cli tool,
but EL6/7's rpm doesn't support file triggers.  Instead, we use a trio
of scriptlets borrowed from remirepo to handle this in a compatible way.